Update pad.py
There was a notable inconsistency in how a singular dataframe without groups worked versus with groups in that the forward filling of names worked for group without the need for an assign, but not for the singular dataframe. 

To resolve this I used a simple trick to find the column that have only a single value extract that name and then forward filled. This removes the need for any additional code as in the example currently in the docs.